% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{cartesian2geographic}
\alias{cartesian2geographic}
\title{cartesian2geographic}
\usage{
cartesian2geographic(x, y = NA, z = NA, useproj4 = FALSE, returnsf = FALSE)
}
\arguments{
\item{x}{numeric, X coordinate OR a vector of length=3 with x, y, z values}

\item{y}{numeric, Y coordinate}

\item{z}{numeric, Z coordinate}
}
\value{
longitude and latitude and ellipsoid height e.g. c(long=3, lat=45.6, h=44)
}
\description{
converts Cartesian coordinates to Spherical coordinates
in WGS84 datum
}
\examples{
cartesian2geographic(-1079348.2174506688, 6121287.922664762, 1425405.2088229598 )
}
