% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_time_interval}
\alias{check_time_interval}
\title{check_time_interval}
\usage{
check_time_interval(interval, units = "secs", tz = "UTC")
}
\arguments{
\item{interval}{numeric value of interval}

\item{units}{string one of "secs", "mins", "hours", "days", "weeks"}

\item{tz}{time zone "" is the current time zone, and "GMT" is UTC}
}
\value{
a "difftime" object
}
\description{
Checks if it is a time interval or creates one from an integer and a unit
}
\details{
check_time_interval
}
\examples{
#
}
