% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{crx2rnx}
\alias{crx2rnx}
\title{Decompress Hatanaka-compressed RINEX format}
\usage{
crx2rnx(filepath)
}
\arguments{
\item{filepath}{path to file that is Hatanaka-compressed RINEX format}
}
\value{
character string of path to decompressed file
}
\description{
Decompress Hatanaka-compressed RINEX format.    

Uses compiled code from crx2rnx.c 
Original C code is from https://terras.gsi.go.jp/ja/crx2rnx.html
Please check license
}
\examples{
ef<-rRINEX::example.files  
crx2rnx(ef$obs2.base)

}
\references{
LICENSE:   
(\href{https://terras.gsi.go.jp/ja/crx2rnx/LICENSE.txt}{https://terras.gsi.go.jp/ja/crx2rnx/LICENSE.txt})
See the following paper for more detail of the compression format and the tools:

PAPER:  
Bulletin of the Geographical Survey Institute, 55, 21-30,
(\href{http://www.gsi.go.jp/ENGLISH/Bulletin55.html}{hhttp://www.gsi.go.jp/ENGLISH/Bulletin55.html})
}
