% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{numeric number}
}
\value{
an integer with the number of decimal places
}
\description{
decimalplaces
}
\examples{
decimalplaces(pi)
decimalplaces(456.345263645)
}
