% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs3.R
\name{decodeEpochObs3}
\alias{decodeEpochObs3}
\title{decodeEpochObs3}
\usage{
decodeEpochObs3(oo, epochHeader, hdr, time, sv, useindicators, verbose = FALSE)
}
\arguments{
\item{oo}{connection to file that is being read}

\item{epochHeader}{header with info from epoch}

\item{hdr}{header with list returning from function "obsheader3"}

\item{time}{}

\item{sv}{}

\item{useindicators}{}

\item{verbose}{}
}
\value{
parsed epoch as matrix with nrow=number of satellites from epochHeader
}
\description{
block processing of each epoch chunk (time step)
}
\examples{
#
}
