% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decompress}
\alias{decompress}
\title{decompress  files}
\usage{
decompress(filepath, verbose = FALSE)
}
\arguments{
\item{filepath}{path of file to be decompressed}

\item{verbose}{boolean, DEFAULT FALSE  to output extra messages}
}
\value{
list of values: 
\itemize{
  \item path - path to decompressed file or, if no compression found, the original 
  \item version - RINEX version
 \item   filetype: "O" or "N"
 \item   rinextype: "obs" or "nav"
 \item   systems: M for multiple, or G for GPS
 G for GPS
 
 R for GLONASS
 
 S for SBAS payload
 
 E for Galileo
 \item   filepath: filepath
  
  }
}
\description{
Decompress .ZIP, .Z or .gz files if file is compressed
}
