% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decZ}
\alias{decZ}
\title{Decompress .Z files (Linux only for now)}
\usage{
decZ(filepath)
}
\arguments{
\item{filepath}{path of file to be decompressed}
}
\value{
path to decompressed file or NULL on problem
}
\description{
Decompress .Z files (Linux only for now)
}
\examples{
decZ(system.file("extdata", "pado348o.20d.Z", package = "rRINEX"))
}
