% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decZip}
\alias{decZip}
\title{Unzip files and return RINEX files in ZIP archive}
\usage{
decZip(filepath)
}
\arguments{
\item{filepath}{path of file to be decompressed}
}
\value{
character vector with names of RINEX files or NULL on error or on missing RINEX files
}
\description{
Unzip files and return a character vector with names of RINEX files
}
