% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs3.R
\name{epoch}
\alias{epoch}
\title{epoch}
\usage{
epoch(data, raw, hdr, time, sv, useindicators, verbose)
}
\arguments{
\item{data}{}

\item{raw}{}

\item{hdr}{}

\item{time}{}

\item{sv}{}

\item{useindicators}{}

\item{verbose}{}
}
\value{
parsed epoch
}
\description{
block processing of each epoch (time step)
}
\examples{
#
}
