% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRINEX-data.R
\docType{data}
\name{example.files}
\alias{example.files}
\title{Example data
--}
\format{
A list with information on some examples RINEX files.
\itemize{
  \item obs.rover exampleRover.20o a RINEX 3.01 format file 
  \item nav       exampleNav.20o a RINEX 3.01 format file 
  \item obs.base  exampleObsBaseStation.20o a RINEX 3.01 format file 
}
}
\usage{
example.files
}
\description{
Some example files
}
\examples{
print(rRINEX::example.files)


}
\keyword{examples}
\keyword{path}
\keyword{to}
