% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{first_nonblank_line}
\alias{first_nonblank_line}
\title{first_nonblank_line   
 Return first non-blank 80 character line in file}
\usage{
first_nonblank_line(f, max_lines = 10)
}
\arguments{
\item{f}{path to file that is Hatanaka-compressed RINEX format}

\item{max_lines, }{INTEGER, maximum number of blank lines}
}
\value{
FIRST NON BLANK LINE IN RINEX
}
\description{
first_nonblank_line   
 Return first non-blank 80 character line in file
}
\examples{
##
}
