% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseStationFiles.R
\name{get.IGS}
\alias{get.IGS}
\title{getFile.liguria}
\usage{
get.IGS(tms, station, onlycheck = F, type = "highrate")
}
\arguments{
\item{tms}{date and time at which the file should be searched for}

\item{station}{name of station of the EUREF Region}

\item{onlycheck}{(default FALSE) do not download the file, just check if it exists and
return TRUE or FALSE}

\item{type}{either "highrate" or other}
}
\value{
file path to downloaded observation and navigation files if download is successful,
or logical FALSE if not, or  if only checking logical TRUE or FALSE
}
\description{
Downloads automatically a base station's observation
and navigation file for a specific time and date. IGS framework stations, see
(\href{https://www.igs.org/maps/}{https://www.igs.org/maps/}).

See (\href{http://www.epncb.oma.be/_networkdata/data_access/highrate/}{http://www.epncb.oma.be/_networkdata/data_access/highrate/})
}
\examples{
## download file from 2 weeks before today from Genova's baset station
## 
filepath<-get.IGS( as.POSIXct(Sys.time() - 14*(60*60*24) ), "GENU" )
print(filepath)
if(is.character(filepath)) file.remove(filepath)
}
