% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_bits}
\alias{get_bits}
\title{Get number of bits in your system architecture}
\usage{
get_bits()
}
\value{
The operating system 64 or 32 bits as integer
}
\description{
Get number of bits in your system architecture
}
