% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{The operating system name}
\usage{
get_os()
}
\value{
The operating system: one of "windows", "linux", "osx" or "solaris"
}
\description{
The operating system name
}
\examples{
get_os()
}
