% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os_info}
\alias{get_os_info}
\title{Get number of bits in your system architecture}
\usage{
get_os_info()
}
\value{
The operating system 64 or 32 bits as text "64bits" or "32bits"
}
\description{
Get number of bits in your system architecture
}
\examples{
get_os_info()
}
