% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoFromRINEX.R
\name{getApproxPositionFromRINEX.OBS.header}
\alias{getApproxPositionFromRINEX.OBS.header}
\title{Get approximate position  from RINEX Observaton file}
\usage{
getApproxPositionFromRINEX.OBS.header(rinexFile, class = "numeric")
}
\arguments{
\item{rinexFile}{observation RINEX file}

\item{class}{set either "numeric" (default), or "sf"   to return these classes.}
}
\value{
latlong values as \code{numeric} c(longitude, latitude) or \code{sf}  objects
}
\description{
Get approximate position ONLY if is in the header of the RINEX observation file
}
\examples{
rinexFile<-rRINEX::example.files[["obs.rover"]]
pos<-getApproxPositionFromRINEX.OBS.header(rinexFile)
print(pos)
pos<-getApproxPositionFromRINEX.OBS.header(rinexFile, "sf")
}
