% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseStationFiles.R
\name{getFile.liguria}
\alias{getFile.liguria}
\title{getFile.liguria}
\usage{
getFile.liguria(timestamp, station, onlycheck = F)
}
\arguments{
\item{timestamp}{date and time at which the file should be searched for}

\item{station}{name of station of the Veneto Region}

\item{onlycheck}{(default FALSE) do not download the file, just check if it exists and
return TRUE or FALSE}
}
\value{
file path to downloaded observation and navigation files if download is successful,
or logical FALSE if not, or  if only checking logical TRUE or FALSE
}
\description{
Downloads automatically a base station's observation
and navigation file for a specific time and date. Liguria Network.

See (\href{http://gnss.regione.liguria.it/}{http://gnss.regione.liguria.it/})
}
\examples{
## download file from 2 weeks before today from Genova's baset station
filepath<-getFile.liguria( as.POSIXct(Sys.time() - 14*(60*60*24) ), "GENU" )
print(filepath)
if(is.character(filepath)) file.remove(filepath)
}
