% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseStationFiles.R
\name{getFile.veneto}
\alias{getFile.veneto}
\title{getFile.veneto}
\usage{
getFile.veneto(timestamp, station)
}
\arguments{
\item{timestamp}{date and time at which the file should be searched for}

\item{station}{name of station of the Veneto Region}
}
\value{
file path to downloaded observation and navigation files
}
\description{
Downloads automatically a base station's observation 
and navigation file for a specific time and date. Veneto Network.   

See (\href{http://retegnssveneto.cisas.unipd.it}{http://retegnssveneto.cisas.unipd.it})
}
