% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{initializeNetCDF}
\alias{initializeNetCDF}
\title{initializeNetCDF}
\usage{
initializeNetCDF(filepath = NA, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{filepath}{optional path to file. If not provided, a temporary file will be created}

\item{overwrite}{boolean value, DFAULT=FALSE}

\item{verbose}{boolean value, DFAULT=FALSE}
}
\value{
boolean FALSE in case of errors, or list with 
\itemize{
\item path2file :path to NetCDF file
\item netcdf: Object of class "NetCDF" which points to the NetCDF dataset    
}
}
\description{
initializeNetCDF
}
\examples{
#
}
