% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{make_time_interval}
\alias{make_time_interval}
\title{make_time_interval}
\usage{
make_time_interval(time1, time2)
}
\arguments{
\item{time1}{POSIXlt date/time object}

\item{time2}{POSIXlt date/time object}
}
\value{
a nanotime::nanoduration object
}
\description{
creates a time interval
}
\details{
make_time_interval
}
\examples{
#
}
