% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs3.R
\name{obsheader3}
\alias{obsheader3}
\title{obsheader3}
\usage{
obsheader3(f, use = NA, meas = NA, verbose = FALSE)
}
\arguments{
\item{f}{file path}

\item{use}{default NA, single character or vector of systems 'G'  or c('G', 'R') or similar}

\item{meas}{default NA, single character or vector of measurments to considre, 
e.g.   'L1C'  or  c('L1C', 'C1C') or similar}

\item{verbose}{boolean, DEFAULT FALSE,    output extra messages}
}
\value{
a list with all header information from RINEX
}
\description{
get RINEX 3 OBS types, for each system type
optionally, select system type and/or measurement type to greatly
speed reading and save memory (RAM, disk)
}
\examples{
#
}
