% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{opener}
\alias{opener}
\title{Decompress Hatanaka-compressed RINEX format}
\usage{
opener(f = NULL, header = FALSE)
}
\arguments{
\item{f}{file path to file that is RINEX format}

\item{header}{read also head in  RINEX format}
}
\value{
ASCII data with RINEX info
}
\description{
Decompress Hatanaka-compressed RINEX format.
}
\examples{
##
}
