% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRINEX-data.R
\docType{data}
\name{paths.to.example.files}
\alias{paths.to.example.files}
\title{Example data
--}
\usage{
paths.to.example.files()
}
\value{
A list with information on some examples RINEX files.
\itemize{
  \item obs.rover exampleRover.20o a RINEX 3.01 format file 
  \item nav       exampleNav.20o a RINEX 3.01 format file 
  \item obs.base  exampleObsBaseStation.20o a RINEX 3.01 format file 
  
}
}
\description{
Some example files
}
\examples{
paths.to.example.files() 


}
\keyword{examples}
\keyword{path}
\keyword{to}
