% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoFromRINEX.R
\name{plotClosestStation}
\alias{plotClosestStation}
\title{Show closest GNSS stations in mapview}
\usage{
plotClosestStation(rinexFile)
}
\arguments{
\item{rinexFile}{observation RINEX file}
}
\value{
Shows mapview of points.
}
\description{
Show closest GNSS stations in mapview
}
