% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseStationFiles.R
\name{plotClosestStations}
\alias{plotClosestStations}
\title{Show closest GNSS stations in mapview}
\usage{
plotClosestStations(rinexFile, nStations = 3, interactive = F)
}
\arguments{
\item{rinexFile}{observation RINEX file}

\item{nStations}{number of closest stations. Default is 3.}

\item{interactive}{Uses \code{tmap} package, sets interactive or static.}
}
\value{
Shows interactive or static map of points.
}
\description{
Show closest GNSS stations in mapview
}
\examples{
 ef<-rRINEX::example.files
 ## plotClosestStation(ef$obs.rover)
}
