% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinex.R
\name{readrnxhApproxPosition}
\alias{readrnxhApproxPosition}
\title{rtklib.readrnxh}
\usage{
readrnxhApproxPosition(f = NULL)
}
\arguments{
\item{f}{RINEX filename}
}
\value{
list with a tue items:
\itemize{
\item APPROX.POSITION.XYZ - Cartesian coordinate in ECEF system of approximate
position
\item APPROX.POSITION.LATLONG - Geographic coordinate in longitude and latitude
}
}
\description{
read RINEX header
}
\examples{
f <- rRINEX::example.files$obs.crx.rover
f <- "data-raw/pado346k.20d.Z"
f <- "data-raw/BZRG00ITA_S_20203461015_15M_01S_MO.rnx"
# rinexobs3(f, verbose=TRUE, tlim1=Sys.time(), tlim2=NA)
}
