% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{rinex_version}
\alias{rinex_version}
\title{rinex_version
 Return RINEX version}
\usage{
rinex_version(s)
}
\arguments{
\item{s}{string with first non blank line of  RINEX file}
}
\value{
A list with information on some examples RINEX files.
\itemize{
  \item version, numeric ,  RINEX/SP3 file version
  \item is_crinex, boolean, is it a Compressed RINEX CRINEX Hatanaka file
}
}
\description{
rinex_version
 Return RINEX version
}
\examples{
##
}
