% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{rinexinfo}
\alias{rinexinfo}
\title{rinexinfo
 Return RINEX version}
\usage{
rinexinfo(f)
}
\arguments{
\item{f}{file connection to RINEX file (not zip/tar/gz/z compressed)}
}
\value{
A list with information on some examples RINEX files.
}
\description{
rinexinfo
 Return RINEX version
}
\examples{
##
}
