% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs3.R
\name{rinexobs3}
\alias{rinexobs3}
\title{rinexobs3}
\usage{
rinexobs3(
  f,
  use = NA,
  tlim1 = NA,
  tlim2 = NA,
  useindicators = FALSE,
  meas = NA,
  verbose = FALSE,
  interval = ".0",
  udt = FALSE
)
}
\arguments{
\item{f}{RINEX OBS 3 filename}

\item{use}{character vector, single value, e.g. 'G'
or many values, e.g. c('G', 'R') see 
[https://files.igs.org/pub/data/format/rinex_clock304.txt](https://files.igs.org/pub/data/format/rinex_clock304.txt)}

\item{tlim1}{POSIXct, read only after this time 
(if tlim2 is provided, must be that  tlim1 < tlim2)}

\item{tlim2}{POSIXct, read only before this time
(if tlim1 is provided, must be that  tlim2 > tlim1)}

\item{useindicators}{SSI, LLI are output}

\item{meas}{'L1C'  or  c('L1C', 'C1C') or similar}

\item{verbose}{boolean, DEFAULT FALSE,    output extra messages}

\item{interval}{allows decimating file read by time e.g. every 5 seconds.
Useful to speed up reading of very large RINEX files}

\item{udt}{boolean, DEFAULT FALSE,   stands for use data.table - this creates and object 
NOT of NetCDF but of a long table with columns - easier to understand for some}
}
\value{
HDF5 array or data.table object if udt=TRUE
}
\description{
from [http://acc.igs.org/misc/rinex304.pdf](http://acc.igs.org/misc/rinex304.pdf)
decodes RINEX 3.04 version files and creates a NetCDF or data.table
}
\examples{
f <- rRINEX::example.files$obs.crx.rover
f <- "data-raw/BZRG00ITA_S_20203461015_15M_01S_MO.rnx"
rinexobs3(f, verbose=TRUE, tlim1=Sys.time(), tlim2=NA)
}
