% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRINEX.R
\docType{data}
\name{rRINEX.RINEX.OBS}
\alias{rRINEX.RINEX.OBS}
\title{rRINEX.RINEX.OBSt}
\format{
An object of class \code{list} of length 2.
}
\usage{
rRINEX.RINEX.OBS
}
\description{
RINEX observation file template object 
like from (\href{https://gage.upc.edu/sites/default/files/gLAB/HTML/Observation_Rinex_v3.01.html}{https://gage.upc.edu/sites/default/files/gLAB/HTML/Observation_Rinex_v3.01.html})
}
\details{
rRINEX.RINEX.OBS
}
\keyword{datasets}
