% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtkget.R
\name{rtkget}
\alias{rtkget}
\title{Download GNSS Data Based on a RINEX Observation File}
\usage{
rtkget(rinex_obs_path, output_dir = "gnss_data")
}
\arguments{
\item{rinex_obs_path}{Character. Path to a RINEX observation file (.obs)}

\item{output_dir}{Character. Directory to save the downloaded files}
}
\value{
A list with downloaded file paths (if successful)
}
\description{
Downloads BRDC navigation, SP3 precise ephemeris, and CLK clock files
for the date found in a given RINEX observation file.
}
