% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtklib.R
\name{rtklib.apply}
\alias{rtklib.apply}
\title{rtklib.checkIsWorking}
\usage{
rtklib.apply(type = "single")
}
\arguments{
\item{type}{- see RTKLIB's RTKPOST options. Files with .conf extension.}
}
\value{
TRUE or FALSE depending if RTKLIB path to rnx2rtkp executable is working
}
\description{
Applies post-processing
}
\examples{
system
}
