% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtklib.R
\name{rtklib.checkIsWorking}
\alias{rtklib.checkIsWorking}
\title{rtklib.checkIsWorking}
\usage{
rtklib.checkIsWorking()
}
\value{
TRUE or FALSE depending if RTKLIB path to rnx2rtkp executable is working
}
\description{
Checks if RTKLIB path to rnx2rtkp executable is working
}
\examples{
rtklib.checkIsWorking()
}
