% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtklib.R
\name{rtklib.rnx2rtkp}
\alias{rtklib.rnx2rtkp}
\title{rnx2rtkp From RTKLIB:
 
 @usage: rnx2rtkp [option]... file file [...]}
\usage{
rtklib.rnx2rtkp(
  k = NA,
  o = "",
  ts,
  te,
  ti = NA,
  p = 2,
  m = 15,
  f = 2,
  v = 3,
  b = F,
  c = NA,
  i = NA,
  h = NA,
  e = NA,
  a = NA,
  n = NA,
  g = NA,
  t = NA,
  u = NA,
  d = 3,
  s = " ",
  r = NA,
  l = NA,
  y = 0,
  x = 0
)
}
\arguments{
\item{k}{file   input options from configuration file [off]}

\item{o}{file   set output file [stdout]}

\item{ts}{ds ts start day/time (ds=y/m/d ts=h:m:s) [obs start time]}

\item{te}{de te end day/time   (de=y/m/d te=h:m:s) [obs end time]}

\item{ti}{tint  time interval (sec) [all]}

\item{p}{mode   mode (0:single, 1:dgps, 2:kinematic, 3:static, 4:moving-base,
5:fixed, 6:ppp-kinematic, 7:ppp-static) [2]}

\item{m}{mask   elevation mask angle (deg) [15]}

\item{f}{freq   number of frequencies for relative mode (1:L1,2:L1+L2,3:L1+L2+L5) [2]}

\item{v}{thres  validation threshold for integer ambiguity (0.0:no AR) [3.0]}

\item{b}{backward solutions [off]}

\item{c}{forward/backward combined solutions [off]}

\item{i}{instantaneous integer ambiguity resolution [off]}

\item{h}{fix and hold for integer ambiguity resolution [off]}

\item{e}{output x/y/z-ecef position [latitude/longitude/height]}

\item{a}{output e/n/u-baseline [latitude/longitude/height]}

\item{n}{output NMEA-0183 GGA sentence [off]}

\item{g}{output latitude/longitude in the form of ddd mm ss.ss' [ddd.ddd]}

\item{t}{output time in the form of yyyy/mm/dd hh:mm:ss.ss [sssss.ss]}

\item{u}{output time in utc [gpst]}

\item{d}{col    number of decimals in time [3]}

\item{s}{sep    field separator [' ']}

\item{r}{x y z  reference (base) receiver ecef pos (m) [average of single pos]
rover receiver ecef pos (m) for fixed or ppp-fixed mode}

\item{l}{lat lon hgt reference (base) receiver latitude/longitude/height (deg/m)
rover latitude/longitude/height for fixed or ppp-fixed mode}

\item{y}{level  output soltion status (0:off,1:states,2:residuals) [0]}

\item{x}{level  debug trace level (0:off) [0]}
}
\value{
TRUE or FALSE depending if RTKLIB path to rnx2rtkp executable is working
}
\description{
Applies post-processing using RTKLIB executable 'rnx2rkp'
}
\details{
Read RINEX OBS/NAV/GNAV/HNAV/CLK, SP3, SBAS message log files and ccompute 
 receiver (rover) positions and output position solutions.
 The first RINEX OBS file shall contain receiver (rover) observations. For the
 relative mode, the second RINEX OBS file shall contain reference
 (base station) receiver observations. At least one RINEX NAV/GNAV/HNAV
 file shall be included in input files. To use SP3 precise ephemeris, specify
 the path in the files. The extension of the SP3 file shall be .sp3 or .eph.
 All of the input file paths can include wild-cards (*). To avoid command
 line deployment of wild-cards, use "..." for paths with wild-cards.
 Command line options are as follows ([]:default). With -k option, the
 processing options are input from the configuration file. In this case,
 command line options precede options in the configuration file.
}
\examples{
system
}
