% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtklib.R
\name{setRTKLIB}
\alias{setRTKLIB}
\title{Sets path to RTLIB}
\usage{
setRTKLIB(RTKLIB_bindir = NULL)
}
\arguments{
\item{RTKLIB_bindir}{path to RTKLIB bin *directory*, i.e. where executable rnx2rtkp is located. 
Default is NULL. If not set (or NULL) then interactively asks user-input.}
}
\value{
logical TRUE on success, or message with error
}
\description{
Sets path to RTLIB
}
\examples{
setRTKLIB("/usr/bin")
}
