% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRINEX-data.R
\docType{data}
\name{stazioniGNSS}
\alias{stazioniGNSS}
\title{Stazioni di Network GNSS (ITA e EUREF)}
\format{
A list with multiple objects of class \code{"sf"} with: 
\itemize{
  \item ITA - stations from (\href{http://retegnssveneto.cisas.unipd.it/gpsitn/}{http://retegnssveneto.cisas.unipd.it/gpsitn/})
  \item IGSNetwork - stations from (\href{https://files.igs.org/pub/station/general/IGSNetwork.csv}{https://files.igs.org/pub/station/general/IGSNetwork.csv})
  }
}
\source{
\href{http://retegnssveneto.cisas.unipd.it/gpsitn}{Rete GNSS Veneto}
}
\usage{
stazioniGNSS
}
\description{
La Rete GNSS Italiana è una rete composta da circa 700 stazioni permanenti 
GNSS diffuse sul territorio nazionale e alcune nazioni limitrofe. 
Con cadenza giornaliera viene eseguito il ricalcolo delle coordinate con il 
software Bernese v. 5.2 e con cadenza settimanale vengono emanate le coordinate 
aggiornate dalla combinazione delle 7 soluzioni giornaliere.
}
\details{
La rete EUREF la trovi  (\href{https://files.igs.org/pub/station/general/IGSNetwork.csv}{https://files.igs.org/pub/station/general/IGSNetwork.csv})
}
\examples{
## NOT RUN: mapview() + mapview(stazioniGNSS$ITA)
## NOT RUN: mapview() + mapview(stazioniGNSS$IGSNetwork)
}
\references{
Rete GNSS Veneto
(\href{http://retegnssveneto.cisas.unipd.it/gpsitn}{Rete GNSS Veneto})
}
\keyword{datasets}
