% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRINEX-data.R
\docType{data}
\name{stazioniGNSSita}
\alias{stazioniGNSSita}
\title{Stazioni della Rete GNSS Italiana}
\format{
An object of class \code{"sf"}.
}
\source{
\href{http://retegnssveneto.cisas.unipd.it/gpsitn}{Rete GNSS Veneto}
}
\usage{
data(stazioniGNSSita)
}
\description{
La Rete GNSS Italiana è una rete composta da circa 700 stazioni permanenti 
GNSS diffuse sul territorio nazionale e alcune nazioni limitrofe. 
Con cadenza giornaliera viene eseguito il ricalcolo delle coordinate con il 
software Bernese v. 5.2 e con cadenza settimanale vengono emanate le coordinate 
aggiornate dalla combinazione delle 7 soluzioni giornaliere.
}
\examples{
data(stazioniGNSSita)  
stazioniGNSSita[1,]
}
\references{
Rete GNSS Veneto
(\href{http://retegnssveneto.cisas.unipd.it/gpsitn}{Rete GNSS Veneto})
}
\keyword{datasets}
