% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseStationFiles.R
\name{timestamp2parts}
\alias{timestamp2parts}
\title{timestamp2parts}
\usage{
timestamp2parts(timestamp)
}
\arguments{
\item{timestamp}{date and time to decompose.}
}
\value{
list with year, yearNc (last two digits), month, day, hour, julianDay
}
\description{
timestamp2parts
}
\examples{
timestamp2parts(Sys.time())
}
