% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_time_interval}
\alias{check_time_interval}
\title{check_time_interval}
\usage{
check_time_interval(interval, units = "secs")
}
\arguments{
\item{interval}{character value, (not numeric to allow sub-microsecond precision)
\itemize{
\item can reach nanosecond precision
by using for example 3.000000008
}}

\item{units}{charact, for now only  "secs" supported, in future also , "mins",
"hours", "days", "weeks"}
}
\value{
a "nanotime::nanoduration" object
}
\description{
Checks if it is a time interval or creates one from an integer and a unit
}
\details{
check_time_interval
}
\examples{
# check_time_interval
}
