% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decGZip}
\alias{decGZip}
\title{decompress Gzip files}
\usage{
decGZip(filepath)
}
\arguments{
\item{filepath}{path of file to be decompressed}
}
\value{
path to decompressed file or NULL on error
}
\description{
decompress Gzip files
}
