% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{decimalsAfterPrecision}
\alias{decimalsAfterPrecision}
\title{decimalplaces}
\usage{
decimalsAfterPrecision(x, n = 0)
}
\arguments{
\item{x}{numeric number}

\item{n}{integer number, default 0}
}
\value{
a character string with the decimals after truncation
}
\description{
decimalplaces
}
\examples{
decimalsAfterPrecision(pi, 0)
decimalsAfterPrecision(pi, 2)
}
