% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decompress}
\alias{decompress}
\title{decompress  files}
\usage{
decompress(filepath, verbose = FALSE)
}
\arguments{
\item{filepath}{path of file to be decompressed}

\item{verbose}{boolean, DEFAULT FALSE  to output extra messages}
}
\value{
path to decompressed file or NULL on error
}
\description{
Decompress .ZIP, .Z or .gz files
}
