% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decompress}
\alias{decompress}
\title{Decompress}
\usage{
decompress(filepath)
}
\arguments{
\item{filepath}{path of file to be decompressed}
}
\value{
path to decompressed file or NULL on error
}
\description{
Decompress .Z or .gz files
}
