% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseStationFiles.R
\name{formatURL}
\alias{formatURL}
\title{formatURL}
\usage{
formatURL(url, tms, stationname = NA)
}
\arguments{
\item{url}{the URL string (see example)}

\item{stationname}{Name of the station e.g. for Padova (Italy) station name is PADO}

\item{timestamp}{time of survey in RINEX file}
}
\value{
url string with substitutions of the following
}
\description{
formatURL
}
\examples{
  
url<-"ftps://gdc.cddis.eosdis.nasa.gov/gps/data/campaign/mgex/hourly/rinex3/\%n/\%h/\%s\%n\%H.\%yd.Z"
formatURL(url, (Sys.time() - 14*60*60*24) )

}
