% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoFromRINEX.R
\name{getClosestStations}
\alias{getClosestStations}
\title{Get closest permenent base station with GNSS}
\usage{
getClosestStations(rinexFile, nStations = 3)
}
\arguments{
\item{rinexFile}{observation RINEX file}

\item{nStations}{number of closest stations to find. Default=3}
}
\value{
An sf object with three closest stations and distances.
}
\description{
Get closest permenent base station with GNSS
}
