% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{opener}
\alias{opener}
\title{Decompress ZIP/GZ/TAR.. and Hatanaka-compressed RINEX format}
\usage{
opener(f = NULL, header = NA, verbose = FALSE)
}
\arguments{
\item{f}{path to  RINEX format file}

\item{header}{list with header info - if not found it will try to read by calling the
corresponding header.}

\item{verbose}{boolean, DEFAULT FALSE  to output extra messages}
}
\value{
A connection in read-only mode to RINEX file with pointer 
right after the header, so it is ready to be read.
}
\description{
Decompress ZIP/GZ/TAR.. and  Hatanaka-compressed RINEX format.
}
\examples{
##
}
