% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{opener}
\alias{opener}
\title{Decompress Hatanaka-compressed RINEX format}
\usage{
opener(f = NULL, header = FALSE, verbose = FALSE)
}
\arguments{
\item{f}{path to file that is RINEX format}

\item{header}{boolean,  DEFAULT FALSE  read also head in  RINEX format}

\item{verbose}{boolean, DEFAULT FALSE  to output extra messages}
}
\value{
A connection in read-only mode to RINEX file
}
\description{
Decompress Hatanaka-compressed RINEX format.
}
\examples{
##
}
