% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinex.R
\name{RINEX}
\alias{RINEX}
\title{Create RINEX object}
\usage{
RINEX(FILE.fp)
}
\arguments{
\item{FILE.fp}{file path to RINEX file.}
}
\description{
Create RINEX object
}
\examples{
rinex.file.path<-system.file("extdata", "example.20o", package = "rRINEX")
rinex.obj<-RINEX(rinex.file.path) 
}
\keyword{RINEX}
