% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRINEX.R
\name{RINEX}
\alias{RINEX}
\title{Create RINEX object}
\usage{
RINEX(filepath)
}
\arguments{
\item{filepath}{file path to RINEX file.}
}
\description{
This class creates a RINEX object thus allows 
to open and process RINEX files. Was tested in Linux, not tested in Windows,
please send feedback
}
\examples{
#rinex.file.path<-system.file("extdata", "example.20o", package = "rRINEX")
#rinex.obj<-RINEX(rinex.file.path) 
}
\keyword{RINEX}
