% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setRTKLIB}
\alias{setRTKLIB}
\title{Sets path to RTLIB}
\usage{
setRTKLIB(RTKLIB_bindir = NULL)
}
\arguments{
\item{RTKLIB_bindir}{path to RTKLIB bin directory. Default is NULL, 
if not set then interactively asks user-input.}
}
\value{
message
}
\description{
Sets path to RTLIB
}
\examples{
setRTKLIB("/usr/bin")
}
